Rewrite hcsshim methods to use generated code
Also simplify how errors are returned so that the Win32 HRESULT can be
extracted. This is an improvement over the current situation, but it
is still problematic in that errors cannot be compared for equality.
We should address this in a future change (it will require removing
the debug info from the errors, which means we may want to move it
to the callers first).