Make go binaries using a private pkg directory
When CGO_ENABLED is set to 0, go needs to rebuild its runtime without
cgo support. This requires write access to GOROOT, which the user often
does not have.

Since this is the non-default configuration, the best thing to do is to
store the pkg cache locally so that Go does not try to overwrite any
existing packages built with CGO_ENABLED=1.

This also requires removing the -N -l flags from go, since building the
go runtime with these flags set often fails.