Support embedding HRESULT codes in errors
This allows the GCS to return HRESULTs to the HCS for certain errors.
HRESULTs are a common way for the HCS to represent errors.

This change only adds HRESULTs to JSON parsing errors at the moment. It
is mostly focused on building the necessary infrastructure. Additional
HRESULTs can be added now as needed.