Tag all files as being for Windows only.
This prevents e.g. go guru when running on a different platform tripping over
these files even when the build tags are correctly set. In my case this was the
case on the github.com/containerd/containerd code base due to hcsshim being
vendored there.

I must confess to a small naughtiness which is that since I was unable to run
mksyscall_windows on my Linux machine I have hand edited both the generator and
the generated file.