Add a utility to configure a network interface in a namespace
Currently, the network namespaces are configured in the root namespace
and then later moved into the new container network namespace. Dealing
with network namespaces in Go is really tricky as one has very little
control over which go routine is run on which kernel thread/process
and the network namespace ('netns.SetNs()') is tied to the kernel
thread/process. Moving the code to a seperate utility allows us to
lock the kernel threads early on and allows us to use the 'netns'
package in a safer way.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>