Use netnscfg to configure the network interface
As pointed out in the previous commit, dealing with network namespaces
in Go is iffy, at best. This commit moves the network configuration
over to use the new utility introduced by the previous commit.

Note, this changes the timing of when the network is configured.
Previously, 'eth0' would be configured in the root network
namespace during 'CreateContainer()' and then later moved to
the runc container network namespace in 'ExecProcess()'. With the
new approach, the network interface is configured directly inside
'ExecProcess()' and 'eth0' remains down until then.

This not only makes the code simpler, but also removes the rather
brittle code, which tried to gather the network configuration in the
root network namespace and then replay it inside the new container
namespace.

/etc/resolv.conf also needed special attention as it is added to
baselayer and with the change of ordering of network config the /etc
must exist before creating the overlay rootfs.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>