Change TerminateProcess to SignalProcess
Now the GCS will receive a SignalProcessOptions field in the
ContainerSignalProcess message (which was renamed from
ContainerTerminateProcess). Rather than sending a SIGTERM to the process
followed by a SIGKILL after a timeout, the GCS will simply send the
specified signal to the process. Renaming has also been done throughout
the code to match this semantic change.