Change permissions of etc directory from 0700 to 0755
This fixes an issue experienced in the ubuntu image where apt-get needs
the /etc directory to be executable to perform an nslookup. When we
switch to using bind mounts specified by docker to get resolv.conf into
the container namespace, we might be able to avoid specifying the
permission bits altogether.