Implements the async bridge loop
Implements the async bridge
Implements the bridge multiplexer
Removes any waiting between commands in favor of a channel writer per request
Fixes the "RegisterExitHook" pattern in favor of a single channel writer on an async goroutine
Temporarily disables the bridge_test suite put puts in place bridge unit tests. These will be moved when the handler funcs are pulled off of the bridge in the second change.