Unplug mapped disks from inside the VM
Currently, when removing mapped disks the disks are unmounted
before the host removes them. The removal from the host is just
a notification and the Linux kernel then performs the clean-up
asynchronously.

If we unplug a larger number of disks this may take some time
and there is no way for the host to know when the clean-up is
finished and new/different disks can be added.

This patch tells the Linux kernel that we are about to unplug
the disks after they were unmounted, allowing it to perform
the clean-up in advance.

This fixes the issue in globalmode where a mkdir fails within
the Linux utility/service VM.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>