Fixes to V2 protocol around CombinedLayers
1. There is a bug in the HCS that is not honoring the Capabilities sent
back via the negotiateProtcol. This is causing the first CreateContainer
call to come in in the protocol v4 which should not be. This had the
unintended consequence that the container cache had an entry for the UVM
and the baseFilesPath for /etc already existed. Both of these are
incorrect in the V2 case. This adds a work around to ignore the first
CreateContainer call and simply return a success case until the HCS is
fixed.
2. Removes the use of the container cache in the V2 hot modify case
because the disks/folders/layers are not associated to a container in
the same way and should not be cleaned up automatically as they can
exist in the lifetime of multiple containers.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>