Move HNS implementation to internal package
This moves the HNS implementation out of the root package into an
internal package, with forwarders for existing clients. Once the
package's interface is updated and stabilized, it can be made external.

This change also eliminates the ContainerHotAdd and ContainerHotRemove
methods, which are not used by any open source software on GitHub. These
methods are redundant with the free functions HotAttachEndpoint and
HotDetachEndpoint.