runhcs: Use random pipe names when communicating
When a container ID is reused, it's possible to get runhcs into a state
where it will try to connect to the wrong VM (e.g. when cleaning up
resources for a container whose host has been deleted and recreated).

Resolve this comprehensively by using a new random ID to generate the
pipe names for each container.