Removing rootfs2vhd use tar2ext4
The command rootfs2vhd was used to create a rootfs.vhd from an OpenGCS
rootfs.tar.gz. In order to use this it was necessary to start a UVM on an
actual machine. The tool tar2ext4 creates a better optimized version of
rootfs.vhd and can do so without a UVM creation. Use this tool instead.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>