Remove tar2vhd from the OpenGCS
tar2vhd is a tool that converted a Docker layer tar to a vhd by use of
streaming the tar to a LCOW UtilityVM and extracting the tar to an ext4
filesystem and streaming back the vhd. This tool has been replaced by a tool
in github.com/Microsoft/hcsshim/cmd/tar2ext4 that can do all of this locally
on a host without the need of a UtilityVM and is much faster. Use this tool
instead moving forward.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>