expose IsOperationAccessIsDenied, IsOperationInvalidState error checks
on RS1, RS2 builds call to GetContainers may fail with 'access denied' or
'invalid state' errors. You may want to have retry loop to resolve this
issue. But there should be a way to check if returned error is one of these
types. This commit exposes IsOperationAccessIsDenied and IsOperationInvalidState
for these cases.

Signed-off-by: Andrey Kolomentsev <andrey.kolomentsev@gmail.com>