Fix bug in runhcs shim with tty=true on close
On close in the case of a tty the caller does not know the close is taking
place and does not have the ability to unblock the upstream pipes. Since the
caller might not be issuing any stdin even though we close the hcs side of the
pipe there is nothing that alerts go to unblock the io.Copy.

This fix stops waiting for Stdin to finish the io.Copy on hcs process exit.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>