Revert removal of WCOW pause container activation
As it turns out on RS5 a Windows Process Isolated container actually
does open the network compartment at the activation of the first
container. So we cannot skip creating the pause container when creating
the pod. This is not true for Windows Hypervisor Isolated which is held
open via the GNS.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>