Forcibly terminate running process after signal
There are cases in Windows where a signal to a process may not actually
cause the process to stop. As a best practice when a signal is expected
to cause the process to stop IE: SIGKILL, SIGTERM we now forcibly
terminate the process after a timeout period if the process does not
stop on its own.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>