Forcibly close stdout, stderr relay after timeout
Implements a safety fallback when the HCS fails to close the stdout, stderr
IO pipes after a process exit. In this case the ioWg would wait forever and
hang the shim and all cleanup logic waiting for the exit event. This code now
forcibly closes the handles if not already closed naturally after the post
process exit timeout.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>