Change HotRemove NIC to warning instead of error
We don't actually fully support HotRemove NIC from namespace but since
this only happens on teardown we likely don't need it. For now just log
this as a warning rather than returning an error because we do
actually try to safe remove it from the host side on shutdown which is
returning an error that is silently ignored.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>