Fix bug that fails to return error when out of memory
Fixed a bug when UVM activation fails that we fail to return the error message
and continue to process the setup code.

Fixed an issue where we nil out the hcsSystem on the UVM after a call to Close
that causes any additional calls to the package to panic the shim rather than
returning the hcs.IsAlreadyClosed(err) error.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>