Add temporary workaround for Windows Process CPU QoS
Windows RS5 has a bug in Windows Process Container CPU QoS that if the
Processor structure is passed at all it assumes the Count variable first and
does not allow setting Limit or Weight. For now silently ignore these
additional QoS features and succeed the activation to run the process.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>