Change HCS TaskExit ownership responsibility
It turns out that eventing the TaskExit at the end of a process is not the
correct time on Windows. In all cases there is a container Silo seperate from
the init process and in Hypervisor isolated cases there is a parent UtilityVM.

This change makes the init process TaskExit notification only fire once the
Silo/UtilityVM are successfully torn down making sure there are no resources in
use when the TaskExit is sent.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>