shimdiag: Do not hang exec -t on Ctrl-Z
os.Stdin.Read() returns EOF if the user presses Ctrl-Z. This makes sense
for non-terminal execs, where we can close stdin for the process and the
process will generally exit. But it does not make sense for terminal
execs, where Ctrl-Z will just cause the process to hang.

When using exec in terminal mode, read raw bytes from the stdin
descriptor to avoid this Ctrl-Z translation.