Separate v1 and v2 bridge interfaces
This change breaks apart the v1 and v2 schema interfaces as well as the v3 and
v4+ bridge protocols. It makes the v4+ protocol opt in so that downlevel
clients running v1 Xenons will stay on v3. This allows us to iterate and change
the model for v4+ without risking breaking downlevel.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>