Simplify the bridge request/return pattern
This change makes it easier for the rquest/response pattern to return a result
or error. This is perferred over the w.Write and w.Error pattern where a
response must be written and then an early return. This makes the code easier
to read and control flow easier to understand.