rootfs: Convert from initrd by extracting and archiving
This fixes files that are hard linked in the root file system.

Before this change, bsdtar was used to convert the cpio initrd archive
to a tar file, but this failed for mkfs* files that are hard linked to
each other. This is arguably a bug in bsdtar, but it is complicated
because tar and cpio store hard linked files in reverse order from each
other (tar stores the payload in the first hard link, while cpio stores
the payload in the last link).