Receive shim options via stdin
containerd has an option struct for our shim which contains both
shim-wide options (e.g. logging level) and task specific options (e.g.
sandbox image to use). Previously containerd only passed these options
in as part of the task creation call, but recently also added support
to pass them in directly to the shim via a protobuf message in stdin.

Now that containerd supports passing the options directly to the shim,
we can use that as the source for shim-wide options, instead of doing
that only when a task is created. This will also make it easier to add
more shim-wide options in the future, as their affect can be applied
immediately on shim start.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>