init: Add option to initialize entropy from vsock
This change adds a command line option, -e, that specifies a vsock port
to read initial RNG entropy. init reads all data written to the
connection by the host, adds this data to the kernel RNG, and increments
the available entropy count. The host is trusted to write
cryptographically random data to this port.

Without this, the available entropy at LCOW boot is very low, and data
read from /dev/urandom is likely to be highly predicable.