Explicitly set namespace when publishing an event
This fixes a regression introduced when we switched to TTRPC for
publishing events to containerd. Previously we explicitly passed the
namespace for each event as a command line parameter to containerd.exe,
which was invoked to publish the event. Now that TTRPC is used, the
context passed to Publish is expected to include the namespace as a
stored value.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>