LCOW UVM metrics support
This change adds the ability to query an LCOW UVM for metrics. The
currently supported metrics are CPU usage in nanoseconds, and working
set size in bytes. Because containerd does not understand the concept of
a UVM, we return a Statistics protobuf that has both a section for
container stats, as well as a section for UVM stats. When the task that
owns the UVM is queried for stats, it will fill out the VM section. In
the future when container stats are supported as well, the container
stats section will also be filled out.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>