Support WCOW single-file mounts
We currently support a temporary fix here to allow WCOW single-file
mounts to function. VSMB does not support directly mapping a single file
into a UVM, but rather supports an `AllowedFileList` along with options
`RestrictFileAccess` and `SingleFileMapping`, which are together used to
ensure that the share only presents a subset of the actual present
files. We map the file into the UVM using this approach, and then use
bindflt in the UVM to present the file in the share at the desired
location in the container file system.

However, naively implementing this introduces some problems. Notably,
the VSMB share map tracks shares by the host path, leading to problems
if you try to map an additional single file from the same directory, or
map a directory as both a set of single files, as well as in another
location as the entire directory.

To fully resolve this issue, we require broader changes to how resources
are managed in hcsshim. But for now we introduce a temporary fix. We now
separately track VSMB shares that present an unrestricted directory, and
those that allow only specific file access. When the VSMB functions
receive a host path, they take the appropriate action based on if that
path points to a  file (map its directory as a share restricted to only
the desired files), or if it points to a directory (map the directory
without restriction).

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>