Support forwarding named pipes to WCOW hypervisor container
Adds support for forwarding a named pipe from the host into a WCOW
hypervisor container (forwarding to a process container appears to
already have been supported). To forward a pipe into a hypervisor
container, the pipe must first be mapped into the UVM via the
MappedPipes resource type. The MappedPipes resource actually just sets
up a VSMB share for the pipes, so in the future we could choose to
manage this concept directly as a VSMB share rather than via the
MappedPipes resource.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>