Fix nil dereference race in Pipe/TtyRelay Wait
The process for creating a V2 container is separated into a few different
steps. First a CreateContainer is called which creates the actual runc
container process with a standard pipe relay but no upstream IO. Next the HCS
sends an ExecProcess call passing in the upstream IO to relay to. If that call
fails, when the HCS calls Kill on the container the relay wait code would panic
because the upstream IO had never been assigned.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>