Fix issue causing SCSI hot add to find no matching device
It appears there is a timing issue in the current code that checks for the
SCSI device to be added. We were waiting for the /block folder to show up but
then immediately read the device entries. There is some amount of time that it
can take for that to show up as well. So this change now waits for both the
/block and /block/sd* entry under the same context timeout.

Signed-off-by: Justin Terry (VM) <juterry@microsoft.com>