Allow a VHD to be mounted into multiple LCOW containers in the same pod
This is a re-visiting of the change made by abwah previously in commit
7f17353012eda1dbd8367135642ec24ebc88bd1d. That work was reverted due to
test failures with WCOW. With this new change, we no longer attempt to
support WCOW (we can look into that later if needed).

This change also differs from the original in that we now generate the
VHD dynamically at test time, rather than checking in a static VHD file.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>