Add test case where container init forks then exits
This adds a new test in the cri-containerd suite. The test runs an LCOW
container with a command line that causes the init process to fork
a background process (which runs forever) then exit. The test then
ensures that the container status is reported as exited, since the init
process has terminated.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>