Fix LCOW SCSI layer mounting
Fixes an issue that broke mounting of SCSI layers. Due to changing the
assignment of the return values from `AddSCSILayer` from `:=` to `=`,
the error value became scoped to the inner if block, which caused the
later return statement to pick up the previous error value instead of
the new one.

This change also improves some error values and adds additional logging.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>