Skip test on Windows before 19H1
The Test_RunPodSandbox_MultipleContainersSameVhd_WCOW test does not work
prior to 19H1. In those versions, HcsFormatWritableLayerVhd requires the
VHD to be mounted prior to calling, which we don't do currently. Because
of this, we are unable to create a VHD to use for the test.

This fixes the issue by simply skipping the test on the affected OS
versions. We could potentially resolve this in the future by adding
more code to do the VHD mount ourselves, but this is a quick fix for
moment.

We will still have test coverage of this functionality on 19H1, so this
shouldn't be a big issue.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>