Add feature flags to cri-containerd tests
This change adds support to the cri-containerd test suite to pass
feature flags on the command line to control what sets of functionality
are tested. The flags are passed as "-feature <feature>", and multiple
flags can be passed to a single invocation. This change also adds a
helper function that can be used to test if a given set of features are
enabled. If any of the required features are not enabled, the current
test is skipped. If no feature flags were passed, the helper function
treats all features as enabled.

This change is needed to support environments where we only want to run
subsets of the tests. For instance, in some cases we don't want to test
LCOW at all.

The -run test command parameter offers similar functionality through
specifying a regex to filter tests. However, it does not seem feasible
to write a regex that could detect the presence of e.g. "LCOW" in both
a top-level test name, or a subtest name. Therefore, this approach is
unsuitable.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>