Add virtual NUMA support
* Added NUMA topology and processor related HCS schema definitions
* New annotations "io.microsoft.virtualmachine.computetopology.numa.defaultnuma"
and "io.microsoft.virtualmachine.computetopology.numa.virtualnodecount" to be
able to specify the number of virtual NUMA nodes to assign the UVM.
* New UVM method hostProcessorInfo that will call HcsGetServiceProperties
and return the hosts processor topology as well as the total number of
logical processors on the system. This is to fix the bug where we were
incorrectly capping a users requested CPU count on > 1 NUMA node systems
as runtime.NumCPU only sees the LPs from the processor/node it is run on.
* Refactored checking for validations/unsupported configurations when creating
the UVM doc.
* Added LCOW test for NUMA validation.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>