Fix UVM processor cap with processor groups.
* Fix bug where we were incorrectly capping a users requested
CPU count in multiple processor group scenarios. This was due to
the fact that we used runtime.NumCPU() to get the logical processor
count of the system and this does not take into account multiple
processor groups. Instead use the processor topology information from HCS
to cap the UVMs amount and the win32 API function GetMaximumProcessorCount
to cap a process isolated containers requested amount.
* Add HCS schema files necessary to retrieve the processor topology information.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>