Add implementation of network configuration proxy
* Ncproxy (abbreviation of network configuration proxy) is a proxy used to facilitate
external configuration of a pods network through a set of TTRPC and GRPC services.
Ncproxy relies on other TTRPC/GRPC services to get the information it needs to perform
its actions.

The full set of services are as follows:

------------------------------------------------------------------------------------------------------

NetworkConfigProxy (TTRPC) - This service is exposed by Ncproxy and is used by the shim.
This service has two rpc methods, RegisterComputeAgent and ConfigureNamespace.
RegisterComputeAgent is used to create a ComputeAgent client in ncproxy mapping the shim
to the network namespace it belongs to for requests later on. ConfigureNamespace is used
to query for the HNS endpoints belonging the namespace passed in so that all the endpoints
belonging to the network we care about can be added.

NetworkConfigAgent (GRPC) - This service is exposed by any application implementing the interface
to the service (/cmd/ncproxy/configagent/configagent.proto). This service has one rpc method,
ConnectNamespaceToNetwork.

NetworkConfigProxy (GRPC) - This service is exposed by Ncproxy and is used by a service implementing
the NetworkConfigAgent GRPC interface. This service has two rpc methods, RegisterNetworkConfigAgent and
AddNIC. RegisterNetworkConfigAgent takes in an address to a NetworkConfigAgent service, creates a client
to the service and then maps it to its network ID that it manages. AddNIC forwards an AddNIC request to
the ComputeAgent service below.

ComputeAgent (TTRPC) - This service is exposed by the shim and is called by ncproxy.
Computeagent has one rpc method, AddNIC. This is called to add a NIC to the shims
Utility VM.

---------------------------------------------------------------------------------------------------------

This is an optional feature that can be enabled by setting the annotation "io.microsoft.network.ncproxy"
and providing an address to a TTRPC service that implements the NetworkConfigProxy TTRPC service defined in
/internal/ncproxyttrpc.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>