Fix RootVpIndex type mistake
* In a minroot configuration, everything assigned to the VM > the hosts
logical processors will return -1 for RootVpIndex, so unmarshalling into a
unsigned integer will fail if the virtual processor count assigned to the UVM is
greater than the hosts amount. This was an oversight as our schema generates
uints as regular ints and when manually changing them I missed that RootVpIndex
was not also a uint.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>