Fix the timestamp bug in pulling a multi layer image.
When pulling a multilayer image, the delta layers can delete some files or create new hard links. We have to do this file deletion or hard link creation after we are done with the ImportLayer call. However, these operations alter the last modification timestamps of the parent directories of these files. Some container applications depend on these timestamps and these applications don't work as expected when we delete the files or create hardlinks. This change reverts the timestamps after doing the delete or hard link creation operation so that the container applications depending on them aren't affected.
Also, until now we only did this for files that aren't symlinks, with this change we will do it for files that are symlinks too. This had to be done because while exporting a layer we sometimes incorrectly adds the symlink attribute to a directory that isn't actually a symlink.

Signed-off-by: Amit Barve <ambarve@microsoft.com>
(cherry picked from commit 945b2201480db3e4836bb819a9b4e9cb94b550d9)
Signed-off-by: Kevin Parsons <kevpar@microsoft.com>