Add prepare-disk command
Prepare-disk command formats a given disk with ext4. This is required for a disk(as passthrough) to be made available within a container. Here is an excerpt from a configuration for such a scenario.

          "mounts": [
            {
              "host_path": "\\.\PHYSICALDRIVE2",
               "container_path": "/disk"
            }
          ]

The gcs/runc combo expects the pass-through disk to be formatted and tries to mount it on a local path within the uvm. Prepare-disk ensures that this restriction is met.

Note: Full disk is formatted with ext4 without any partitioning. TODO: add options to allow partitioning and then formatting.

Signed-off-by: Krishnakumar R(KK) <65895020+kk-src@users.noreply.github.com>