Write 1 to memory.use_hierarchy to enable hierarchy support
`Created nested cgroup for controller "memory" which has incomplete hierarchy support. Nested cgroups may change behavior in the future.
cgroup: "memory" requires setting use_hierarchy to 1 on the root.`

After booting the UVM, this is present in dmesg. There's numerous threads on this over the years from docker, lxc and many others. This should enable the hierarchy support we (believe) we're using for our nested container cgroups so the resource consumptions are properly propagated upwards to their parent cgroup (/containers).`

* Enable hierarchy support so memory usage and limits can flow upwards to the parent cgroup like we'd expect for the /container and its nested cgroups. It's not necessary to set this on any nested cgroups, only the root.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>