Allow passing any uid for container processes.
Usually if a username is provided when starting a process inside the container
we look inside the /etc/passwd file of the container to find the uid and gid for that
user. However, if a uid is provided instead of a username there is no need to look into
the /etc/passwd file to see if that user exists.

Signed-off-by: Amit Barve <ambarve@microsoft.com>