Add teleport service, loopback device and cifs mount structures for LCOW
To support teleportable images we need to go about layer setup a bit differently.
The flow here is to first cifs mount the share inside of the UVM itself, and then
for each block device (layer vhd) mount it loopback. After these steps the flow
is the exact same as usual, just supply the paths to the layers from within the UVM
and then we can union them with overlayfs.

We first need some way to get the credentials for the share itself which will come
from the teleport daemon on a teleport enabled host. The daemon hosts a grpc service with a cache
of this information that we can use to retrieve them by providing a key.

* Add LoopbackDevice structure to represent a loopback mount/device in the guest.
* Add CifsMount structure to represent a cifs mount inside the guest.
* Add logic to mount teleportable images
* Add teleportd ManifestStore service grpc definitions

Signed-off-by: Daniel Canter <dcanter@microsoft.com>