Make directory creation deterministic
Prior to this commit, the layout of created ext4 filesystems created using tar2ext4 wasn't
deterministic. Recursive directory creation would change from run to run due to the usage
of a map as temporary storage for items from the tar stream.

This commit changes recursive directory creation to follow the same deterministic pattern
as writeDirectory.

At this point, the only possible source of non-determinism in a create file system is the
UUID in the VHD footer. If no VHD footer is included, then the results are deterministic.