Remove netnscfg gcstools utility
This utility only existed because go used to not play well with network namespaces
due to it's nature of goroutine thread multiplexing. This poor behavior was remedied/solved in go
1.10 due to some changes in how runtime.LockOSThread works with regards to scheduling new goroutines
and spawning new threads.

Great article on the problem: https://www.weave.works/blog/linux-namespaces-and-go-don-t-mix

and the eventual resolution: https://www.weave.works/blog/linux-namespaces-golang-followup

* Add DoInNetNS function to perform a function in a specific network namespace.
* Convert netnscfg to a function to be used in tandem with DoInNetNS

Signed-off-by: Daniel Canter <dcanter@microsoft.com>