Add -endpoint-address flag to cri-containerd tests
This test flag allows control over what address we use to talk to CRI.
Using this allows testing with upstream containerd that only supports
named pipe.

This change also takes a dependency on a Kubernetes utility function to
resolve the address to a dialer of the appropriate type. This brings in
some grossness in test/go.mod due to the way k8s handles their modules,
but I don't think it's a big deal given this is test code.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>