Set CREATE_BREAKAWAY_FROM_JOB flag for job container processes
We don't want to inherit the job object of whatever process is running the job container code (the containerd-shim
generally but this would apply for any process). Set the CREATE_BREAKAWAY_FROM_JOB flag on job container processes
to prevent this from happening. The job object itself will also need to have the JOB_OBJECT_LIMIT_BREAKAWAY_OK limit
set for this to take affect.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>