test: remove k8s.io/kubernetes dependency by using containerd copy
The k8s.io/kubernetes  dependency is only needed for a single function
(GetAddressAndDialer), which doesn't seem to be in any module, other than
k8s.io/k8s itself.

Containerd created a copy of this utility for that reason, so let's use that
copy to get rid of the dependency on k8s.io/k8s.

Perhaps we should try to have that utils package included in one of the
smaller k8s.io moduless.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>