Edit forked version of os/exec package to expose ProcThreadAttrList
The go stdlib doesn't expose a way to set any fields on the STARTUPINFOEX structure that can be used
in the CreateProcess family of calls. It's not so much that we need that struct exposed, but rather we'd
like the ability to set attributes on the PROC_THREAD_ATTRIBUTE_LIST struct that lives in STARTUPINFOEX.
This change adds a forked version of os/exec that lets us pass in what we need. The functionality is exposed
through a new field on SysProcAttr which allows us to take in a *windows.ProcThreadAttributeList structure.

This allows us a couple things, one of which being able to launch a process in a job object at start time instead
of assigning the process after it's already up and running. This is ideal in the case of job containers as (even though
a small timeframe) a process we launch may finish executing before we assign it to the job. This also opens the door
to using the pseudo console (CreatePseudoConsole and friends) api.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>