add support for dm verity target when mounting VPMem devices
Update VPMem mount APIs to support passing dm verity information
in addition to VPMem multi-mapping.

Hash device is expected to be the same as the data device with
hash tree appended right after the ext4 file system data, block
sizes for data and hash devices are expected to be the same as
well.

Additionally handle a case when both multi-mapping and verity
are enabled, in that case, first create dm-linear target and use
that target as a data and hash device for dm-verity target.

Signed-off-by: Maksim An <maksiman@microsoft.com>