Fix stderr comment in containerd-shim serve command
In a comment in the serve command for the containerd-shim it says that to signify that we're
successfully serving the ttrpc endpoint we should close stderr, but we actually close stdout
to signify this. The write end of the pipe is hooked up to stdout on the "serve" command invocation
of the shim and we simply forward stderr to the read side until close. Stderr for the serve
invocation is hooked up to the panic.log file in case the shim panics.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>