Minor fixes to shim panic log & create task functions
The change to collect shim panic logs during shim delete command does not work in cases
when the delete command itself runs into some error. To avoid losing shim panic logs in
such cases we log the shim panic logs (if found) as first thing in the delete command.

CreateTask function had `wcl` mutex lock that wasn't really being used anywhere, this
change removes that. We also don't add a `nil` entry for a new task in the `workloadTasks`
map anymore to avoid the shim panic in some cases where a `GetTask` function might
be called while we are still in the process of creating the task and haven't updated the
nil entry with the actual task struct reference.