Add new flags to integration tests to specify virtstack
Add some new flags to the integration test suite that enables choosing what virtstack
to use for hypervisor isolated containers. This makes it so we can re-use our existing tests
and just pass in new flags to run them with a new stack.

To have this play nice with the test suite I changed `getRunPodSandboxRequest` to take in whatever
annotations we'd want to set on the pod config directly instead of us having to set them on the
returned object itself.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>