internal/cmd: Close individual IO pipes when the relay finishes
The shim is expected to close its end of the IO pipes from the gcs when
it is done using them. This is done to ensure that no data is left
buffered in the pipes on the gcs's end. Previously, this was
accomplished via the ioChannel closing its underlying connection if Read
returned EOF.

However, this is not sufficiently robust, as it will not work in cases
where the shim's IO relay breaks on the write end (e.g. if CRI has gone
away).

To resolve this, we now expose individual methods on cow.Process to
close each IO pipe (in/out/err), and call those from the Cmd
implementation once the IO relay completes.

This should be a good first-pass fix here, until we can apply some more
focused cleanup to the IO relay code in the future.

Some minor renaming/cleanup as well.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>