Remove internal GCS connection functionality
HCS maintains an internal guest connection to the GCS normally if you request it.
However, there are certain features that require us to maintain an external connection
(external in this sense meaning not in HCS) instead like late cloning.

We had swapped to always managing the connection to the GCS ourselves some time ago and
afaik there's been no fallout from it, so I propose let's get rid of the internal branches
altogether. This greatly simplifies the work for going through a different virtstack for
hypervisor isolated containers as well.

Ran go mod vendor in /test to bring in the changes as well.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>