Add support to encrypt SCSI scratch disks with dm-crypt
This protects the files generated by the guest from the host OS, as they
are encrypted by a key that the host doesn't know.

This commit adds a new argument to the scsi.Mount() function, `encrypted`,
that makes the SCSI drive be mounted using dm-crypt. It also uses
dm-integrity for integrity checking. This makes the boot process a couple
of seconds slower.

Also, it adds scsi.Unmount(), which also has the `encrypted` argument,
and it does the necessary cleanup for a drive that has been mounted as
an encrypted drive.

All the pre-existing SCSI tests have been fixed to work with the new
scsi.Mount() function prototype. New tests have been added for the new
code.

This is all disabled for now, it has to be enabled in a future patch.

Important note: This depends on cryptsetup and mkfs.ext4. Also, the
kernel must be compiled with dm-crypt and dm-integrity support.